+++ R/facet_color.R
@@ -2,9 +2,10 @@
 #'
 #' `facet_grid_color` behaves similarly to [ggplot2::facet_grid()] in that it
 #' forms a matrix of panels defined by row and column faceting variables. The
-#' main difference is that it also allows the user to specify the background
-#' colors of the individual facet strips using the `colors` argument. If you
-#' have only one variable with many levels, try [facet_wrap_color()].
+#' main difference is that it also allows the user to specify the background and
+#' label colors of the individual facet strips using the `colors` and
+#' `lab_colors` arguments. If you have only one variable with many levels, try
+#' [facet_wrap_color()].
 #'
 #' @param colors Specifies which colors to use to replace the strip backgrounds.
 #'   Either A) a function that returns a color for a given strip label, B) the
@@ -26,6 +27,7 @@
 #' @inheritParams ggplot2::facet_grid
 #' @importFrom ggplot2 ggproto FacetGrid ggproto_parent
 #' @importFrom rlang arg_match0 is_function
+#' @family facetting functions
 #' @export
 #'
 #' @examples
@@ -181,7 +183,6 @@ convert_lab_colors <- function(lab_colors) {
 #' @export
 #' @importFrom ggplot2 ggproto FacetGrid ggproto_parent
 FacetGridColor <- ggproto("FacetGridColor", FacetGrid,
-  shrink = TRUE,
   draw_panels = function(panels, layout, x_scales, y_scales, ranges, coord,
                          data, theme, params, self) {
     panel_table <-
@@ -216,10 +217,10 @@ FacetGridColor <- ggproto("FacetGridColor", FacetGrid,
 #'
 #' `facet_wrap_color` behaves similarly to [ggplot2::facet_wrap()] in that it
 #' wraps a 1d sequence of panels into 2d. The main difference is that it also
-#' allows the user to specify the background colors of the individual facet
-#' strips using the `colors` argument. This is generally a better use of screen
-#' space than [facet_grid_color()] because most displays are roughly
-#' rectangular.
+#' allows the user to specify the background and label colors of the individual
+#' facet strips using the `colors` and `lab_colors` arguments. This is generally
+#' a better use of screen space than [facet_grid_color()] because most displays
+#' are roughly rectangular.
 #'
 #' @param colors Specifies which colors to use to replace the strip backgrounds.
 #'   Either A) a function that returns a color for a given strip label, B) the
@@ -241,6 +242,7 @@ FacetGridColor <- ggproto("FacetGridColor", FacetGrid,
 #' @inheritParams ggplot2::facet_wrap
 #' @importFrom ggplot2 ggproto FacetWrap ggproto_parent
 #' @importFrom rlang arg_match0
+#' @family facetting functions
 #' @export
 #'
 #' @examples
@@ -338,13 +340,268 @@ wrap_as_facets_list <- function(...) {
 #' @importFrom ggplot2 ggproto FacetWrap ggproto_parent
 #' @importFrom rlang is_function
 FacetWrapColor <- ggproto("FacetWrapColor", FacetWrap,
-  shrink = TRUE,
   draw_panels = function(panels, layout, x_scales, y_scales, ranges, coord,
                          data, theme, params, self) {
     panel_table <-
       ggproto_parent(FacetWrap, self)$draw_panels(panels, layout,
                                                   x_scales, y_scales, ranges,
                                                   coord, data, theme, params)
+    strips <- grep("strip", panel_table$layout$name)
+    for (i in strips) {
+      label <-
+        panel_table$grobs[[i]]$grobs[[1]]$children[[2]]$children[[1]]$label
+      fill <- tryCatch(params$colors(label), error = function(e) NA)
+      if (!is.na(fill)) {
+        panel_table$grobs[[i]]$grobs[[1]]$children[[1]]$gp$fill <- fill
+        if (!is_function(params$lab_colors) && params$lab_colors == "auto") {
+          panel_table$grobs[[i]]$grobs[[1]]$children[[2]]$children[[1]]$
+            gp$col <- white_or_black(fill)
+        }
+      }
+      if (is_function(params$lab_colors)) {
+        color <- tryCatch(params$lab_colors(label), error = function(e) NA)
+        if (!is.na(color)) {
+          panel_table$grobs[[i]]$grobs[[1]]$children[[2]]$children[[1]]$
+            gp$col <- color
+        }
+      }
+    }
+    panel_table
+  }
+)
+
+new_grid_facets <- function(...) {
+  asNamespace("ggh4x")$new_grid_facets(...)
+}
+
+assert_strip <- function(...) {
+  asNamespace("ggh4x")$assert_strip(...)
+}
+
+#' Layout panels in a grid with nested colored strips
+#'
+#' `facet_nested_color` behaves similarly to [ggh4x::facet_nested()] in that it
+#' forms a matrix of panels defined by row and column faceting variables and
+#' nests grouped facets. The main difference is that it also allows the user to
+#' specify the background and label colors of the individual facet strips using
+#' the `colors` and `lab_colors` arguments.
+#'
+#' @inheritParams ggh4x::facet_nested
+#' @inheritParams facet_grid_color
+#' @inherit ggh4x::facet_nested details
+#' @importFrom ggplot2 element_line element_blank
+#' @importFrom ggh4x strip_nested
+#' @importFrom grid unit
+#' @family facetting functions
+#' @export
+#' @examples
+#' df <- data.frame(x = 1:10, y = 1:10,
+#'                  period = factor(c("Permian", "Triassic", "Jurassic",
+#'                                    "Cretaceous", "Paleogene"),
+#'                                  levels = c("Permian", "Triassic",
+#'                                             "Jurassic", "Cretaceous",
+#'                                             "Paleogene")),
+#'                  era = factor(c("Paleozoic", "Mesozoic", "Mesozoic",
+#'                                 "Mesozoic", "Cenozoic"),
+#'                                 levels = c("Paleozoic", "Mesozoic",
+#'                                            "Cenozoic")))
+#' ggplot(df, aes(x, y)) +
+#'   geom_point() +
+#'   facet_nested_color(~ era + period, colors = rbind(periods, eras))
+facet_nested_color <- function(rows = NULL, cols = NULL, scales = "fixed",
+                               space = "fixed", axes = "margins",
+                               remove_labels = "none", independent = "none",
+                               shrink = TRUE, labeller = "label_value",
+                               colors = stages, lab_colors = "auto",
+                               as.table = TRUE, switch = NULL, drop = TRUE,
+                               margins = FALSE,
+                               nest_line = element_line(inherit.blank = TRUE),
+                               solo_line = FALSE, resect = unit(0, "mm"),
+                               render_empty = TRUE,
+                               strip = strip_nested(), bleed = NULL) {
+  colors <- convert_colors(colors)
+  lab_colors <- convert_lab_colors(lab_colors)
+
+  # copied from ggh4x v0.2.8
+  strip <- assert_strip(strip)
+  if (!is.null(bleed)) {
+    lifecycle::deprecate_warn(
+      when = "0.2.0",
+      what = "facet_nested(bleed)",
+      details = paste0("The `bleed` argument should be set in the ",
+                       "`strip_nested()` function instead.")
+    )
+    strip$params$bleed <- isTRUE(bleed)
+  }
+  # Convert logical to elements for backward compatibility
+  if (isTRUE(nest_line)) {
+    nest_line <- element_line()
+  }
+  if (isFALSE(nest_line)) {
+    nest_line <- element_blank()
+  }
+  if (!inherits(nest_line, c("element_line", "element_blank"))) {
+    cli::cli_abort(
+      "The {.arg nest_line} argument must be {.cls element_blank} or inherit \\
+      from {.cls element_line}."
+    )
+  }
+
+  params <- list(nest_line = nest_line, solo_line = isTRUE(solo_line),
+                 resect = resect, colors = colors, lab_colors = lab_colors)
+
+  new_grid_facets(
+    rows, cols,
+    scales, space, axes, remove_labels, independent,
+    shrink, labeller, as.table, switch,
+    drop, margins, render_empty, strip,
+    params = params,
+    super = FacetNestedColor
+  )
+}
+
+#' @rdname facet_nested_color
+#' @format NULL
+#' @usage NULL
+#' @export
+#' @importFrom ggplot2 ggproto ggproto_parent
+#' @importFrom ggh4x FacetNested
+#' @importFrom rlang is_function
+FacetNestedColor <- ggproto("FacetNestedColor", FacetNested,
+  draw_panels = function(panels, layout, x_scales, y_scales, ranges, coord,
+                         data, theme, params, self) {
+    panel_table <-
+      ggproto_parent(FacetNested, self)$draw_panels(panels, layout,
+                                                    x_scales, y_scales, ranges,
+                                                    coord, data, theme, params)
+    strips <- grep("strip", panel_table$layout$name)
+    for (i in strips) {
+      label <-
+        panel_table$grobs[[i]]$grobs[[1]]$children[[2]]$children[[1]]$label
+      fill <- tryCatch(params$colors(label), error = function(e) NA)
+      if (!is.na(fill)) {
+        panel_table$grobs[[i]]$grobs[[1]]$children[[1]]$gp$fill <- fill
+        if (!is_function(params$lab_colors) && params$lab_colors == "auto") {
+          panel_table$grobs[[i]]$grobs[[1]]$children[[2]]$children[[1]]$
+            gp$col <- white_or_black(fill)
+        }
+      }
+      if (is_function(params$lab_colors)) {
+        color <- tryCatch(params$lab_colors(label), error = function(e) NA)
+        if (!is.na(color)) {
+          panel_table$grobs[[i]]$grobs[[1]]$children[[2]]$children[[1]]$
+            gp$col <- color
+        }
+      }
+    }
+    panel_table
+  }
+)
+
+new_wrap_facets <- function(...) {
+  asNamespace("ggh4x")$new_wrap_facets(...)
+}
+
+#' Ribbon of panels with nested colored strips
+#'
+#' `facet_nested_wrap_color` behaves similarly to [ggh4x::facet_nested_wrap()]
+#' in that it wraps a sequence of panels onto a two-dimensional layout, and
+#' nests grouped facets where possible.. The main difference is that it also
+#' allows the user to specify the background and label colors of the individual
+#' facet strips using the `colors` and `lab_colors` arguments.
+#'
+#' @inheritParams ggh4x::facet_nested_wrap
+#' @inheritParams facet_wrap_color
+#' @inherit ggh4x::facet_nested_wrap details
+#' @importFrom ggplot2 element_line element_blank
+#' @importFrom ggh4x strip_nested
+#' @importFrom grid unit
+#' @family facetting functions
+#' @export
+#' @examples
+#' df <- data.frame(x = 1:10, y = 1:10,
+#'                  period = factor(c("Permian", "Triassic", "Jurassic",
+#'                                    "Cretaceous", "Paleogene"),
+#'                                  levels = c("Permian", "Triassic",
+#'                                             "Jurassic", "Cretaceous",
+#'                                             "Paleogene")),
+#'                  era = factor(c("Paleozoic", "Mesozoic", "Mesozoic",
+#'                                 "Mesozoic", "Cenozoic"),
+#'                                 levels = c("Paleozoic", "Mesozoic",
+#'                                            "Cenozoic")))
+#' ggplot(df, aes(x, y)) +
+#'   geom_point() +
+#'   facet_nested_wrap_color(~ era + period, colors = rbind(periods, eras))
+facet_nested_wrap_color <- function(
+    facets, nrow = NULL, ncol = NULL,
+    scales = "fixed", axes = "margins",
+    remove_labels = "none",
+    shrink = TRUE, labeller = "label_value",
+    colors = stages, lab_colors = "auto",
+    as.table = TRUE, drop = TRUE,
+    dir = "h", strip.position = "top",
+    nest_line = element_line(inherit.blank = TRUE),
+    solo_line = FALSE,
+    resect = unit(0, "mm"),
+    trim_blank = TRUE,
+    strip = strip_nested(),
+    bleed = NULL
+) {
+  colors <- convert_colors(colors)
+  lab_colors <- convert_lab_colors(lab_colors)
+
+  # copied from ggh4x v0.2.8
+  strip <- assert_strip(strip)
+  if (!is.null(bleed)) {
+    lifecycle::deprecate_warn(
+      when = "0.2.0",
+      what = "facet_nested_wrap(bleed)",
+      details = paste0("The `bleed` argument should be set in the ",
+                       "`strip_nested()` function instead.")
+    )
+    strip$params$bleed <- isTRUE(bleed)
+  }
+  # Convert logical to elements for backward compatibility
+  if (isTRUE(nest_line)) {
+    nest_line <- element_line()
+  }
+  if (isFALSE(nest_line)) {
+    nest_line <- element_blank()
+  }
+  if (!inherits(nest_line, c("element_line", "element_blank"))) {
+    cli::cli_abort(
+      "The {.arg nest_line} argument must be {.cls element_blank} or inherit \\
+      from {.cls element_line}."
+    )
+  }
+  params <- list(nest_line = nest_line, solo_line = isTRUE(solo_line),
+                 resect = resect, colors = colors, lab_colors = lab_colors)
+  new_wrap_facets(
+    facets, nrow, ncol,
+    scales, axes, remove_labels,
+    shrink, labeller,
+    as.table, drop, dir,
+    strip.position, strip,
+    trim_blank, params,
+    super = FacetNestedWrapColor
+  )
+}
+
+#' @rdname facet_nested_wrap_color
+#' @format NULL
+#' @usage NULL
+#' @export
+#' @importFrom ggplot2 ggproto ggproto_parent
+#' @importFrom ggh4x FacetNestedWrap
+#' @importFrom rlang is_function
+FacetNestedWrapColor <- ggproto("FacetNestedWrapColor", FacetNestedWrap,
+  draw_panels = function(panels, layout, x_scales, y_scales, ranges, coord,
+                         data, theme, params, self) {
+    panel_table <-
+      ggproto_parent(FacetNestedWrap, self)$draw_panels(panels, layout,
+                                                        x_scales, y_scales,
+                                                        ranges, coord, data,
+                                                        theme, params)
     strips <- grep("strip", panel_table$layout$name)
     for (i in strips) {
       label <-